+++ internal/entities/reservation_request.go
@@ -13,7 +13,7 @@ type ReservationRequest struct {
 	Status          string    `json:"status"`
 	StartTime       time.Time `json:"start_time"`
 	EndTime         time.Time `json:"end_time"`
-	TotalPrice      float64   `json:"total_price"`
+	TotalPrice      float32   `json:"total_price"`
 	Language        string    `json:"language"`
 }
 
@@ -36,5 +36,5 @@ type ReservationResponse struct {
 	EndTime           time.Time `json:"end_time"`
 	CreatedAt         time.Time `json:"created_at"`
 	UpdatedAt         time.Time `json:"updated_at"`
-	TotalPrice        float64   `json:"total_price,omitempty"`
+	TotalPrice        float32   `json:"total_price,omitempty"`
 }
